+++ clang/test/Sema/attr-noinline.cpp
@@ -25,3 +25,16 @@ void foo() {
 }
 
 [[clang::noinline]] static int i = bar(); // expected-warning {{'noinline' attribute only applies to functions and statements}}
+
+// This used to crash the compiler.
+template<int D>
+int foo(int x) {
+  [[clang::noinline]] return foo<D-1>(x + 1);
+}
+
+// FIXME: This should warn that noinline statement attribute has higher
+// precedence than the always_inline function attribute.
+template<int D> [[clang::always_inline]]
+int bar(int x) {
+  [[clang::noinline]] return bar<D-1>(x + 1);
+}